# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Petrolex API",
        "description": "Documentation API pour Petrolex",
        "contact": {
            "email": "test@example.com"
        },
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8001",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/orders/payment-callback": {
            "post": {
                "tags": [
                    "Commandes"
                ],
                "summary": "Callback de paiement d'une commande",
                "description": "Traite le callback de paiement envoyé par le prestataire de paiement.",
                "operationId": "api.orders.paymentCallback",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/OrderPaymentCallbackRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Callback traité avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/OrderPaymentCallbackResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erreur de validation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "OrderPaymentCallbackRequest": {
                "required": [
                    "order_id",
                    "status",
                    "transaction_reference"
                ],
                "properties": {
                    "order_id": {
                        "description": "ID de la commande",
                        "type": "integer",
                        "example": 123
                    },
                    "status": {
                        "description": "Statut du paiement (paid, failed, pending, etc.)",
                        "type": "string",
                        "example": "paid"
                    },
                    "transaction_reference": {
                        "description": "Référence de la transaction",
                        "type": "string",
                        "example": "TXN-123456789"
                    },
                    "payment_method": {
                        "description": "Méthode de paiement",
                        "type": "string",
                        "example": "mobile_money"
                    },
                    "amount": {
                        "description": "Montant payé",
                        "type": "number",
                        "format": "float",
                        "example": 23000
                    },
                    "meta": {
                        "description": "Données additionnelles du callback",
                        "type": "object",
                        "additionalProperties": {
                            "type": "string"
                        }
                    }
                },
                "type": "object"
            },
            "OrderPaymentCallbackResponse": {
                "properties": {
                    "_metadata": {
                        "properties": {
                            "success": {
                                "type": "boolean",
                                "example": true
                            },
                            "message": {
                                "type": "string",
                                "example": "Callback de paiement traité avec succès"
                            }
                        },
                        "type": "object"
                    },
                    "data": {
                        "properties": {
                            "order_id": {
                                "type": "integer",
                                "example": 123
                            },
                            "status": {
                                "type": "string",
                                "example": "paid"
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
...
            "ErrorResponse": {
                "properties": {
                    "_metadata": {
                        "properties": {
                            "success": {
                                "type": "boolean",
                                "example": false
                            },
                            "message": {
                                "type": "string",
                                "example": "Une erreur est survenue."
                            }
                        },
                        "type": "object"
                    },
                    "data": {
                        "type": "object",
                        "example": null,
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}